only generate udev rules if device isn't official

+++ setup.sh
@@ -20,18 +20,20 @@ mkdir /r;
 # mount manjaro rootfs
 mount /data/rootfs.img /r;
 
-# mount android gsi
-mount /r/var/lib/lxc/android/android-rootfs.img /s
+if [ ! -f "/r/usr/lib/manjaro-libhybris/no-udev" ]; then
+   # mount android gsi
+   mount /r/var/lib/lxc/android/android-rootfs.img /s
 
-# Set udev rules
-ui_print "Setting udev rules";
-cat /s/ueventd*.rc /vendor/ueventd*.rc | grep ^/dev | sed -e 's/^\/dev\///' | awk '{printf "ACTION==\"add\", KERNEL==\"%s\", OWNER=\"%s\", GROUP=\"%s\", MODE=\"%s\"\n",$1,$3,$4,$2}' | sed -e 's/\r//' > /data/70-manjaro.rules;
+   # Set udev rules
+   ui_print "Setting udev rules";
+   cat /s/ueventd*.rc /vendor/ueventd*.rc | grep ^/dev | sed -e 's/^\/dev\///' | awk '{printf "ACTION==\"add\", KERNEL==\"%s\", OWNER=\"%s\", GROUP=\"%s\", MODE=\"%s\"\n",$1,$3,$4,$2}' | sed -e 's/\r//' > /data/70-manjaro.rules;
 
-# umount android gsi
-umount /s;
+   # umount android gsi
+   umount /s;
 
-# move udev rules inside rootfs
-mv /data/70-manjaro.rules /r/etc/udev/rules.d/70-$VENDOR_DEVICE_PROP.rules;
+   # move udev rules inside rootfs
+   mv /data/70-manjaro.rules /r/etc/udev/rules.d/70-$VENDOR_DEVICE_PROP.rules;
+fi
 
 # If we should flash the kernel, do it
 if [ -e "/r/boot/boot.img" ]; then
